+++ src/Zafiro.FileSystem/Change.cs
@@ -1,0 +1,10 @@
+﻿namespace Zafiro.FileSystem;
+
+public enum Change
+{
+    FileCreated,
+    FileDeleted,
+    DirectoryCreated,
+    FileContentsChanged,
+    DirectoryDeleted
+}
