+++ src/Zafiro.FileSystem/FileSystemChange.cs
@@ -1,0 +1,3 @@
+﻿namespace Zafiro.FileSystem;
+
+public record FileSystemChange(ZafiroPath Path, Change Change);
