+++ src/Zafiro.FileSystem/Zafiro.FileSystem.csproj
@@ -8,12 +8,10 @@
   <ItemGroup>
     <PackageReference Include="CSharpFunctionalExtensions" Version="2.40.0" />
     <PackageReference Include="JetBrains.Annotations" Version="2022.3.1" />
-    <PackageReference Include="morelinq" Version="3.3.2" />
+    <PackageReference Include="morelinq" Version="4.1.0" />
     <PackageReference Include="Serilog" Version="2.10.0" />
     <PackageReference Include="System.IO.Abstractions" Version="19.2.69" />
     <PackageReference Include="System.Reactive" Version="6.0.0" />
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\Zafiro\src\Zafiro\Zafiro.csproj" />
+    <PackageReference Include="Zafiro" Version="4.0.22" />
   </ItemGroup>
 </Project>
