+++ test/Zafiro.FileSystem.Tests/HashCompareStrategyTests.cs
@@ -76,6 +76,7 @@ public class HashCompareStrategyTests
 
         public Task<Result> SetContents(IObservable<byte> contents, CancellationToken cancellationToken = default) => throw new NotImplementedException();
         public Task<Result<Stream>> GetData() => throw new NotImplementedException();
+        public Task<Result> SetData(Stream stream, CancellationToken cancellationToken = default) => throw new NotImplementedException();
 
         public Task<Result> SetData(Stream data) => throw new NotImplementedException();
     }
