+++ test/Zafiro.FileSystem.Tests/SizeCompareStrategy.cs
@@ -53,6 +53,7 @@ public class SizeCompareStrategyTests
 
         public Task<Result> SetContents(IObservable<byte> contents, CancellationToken cancellationToken = default) => throw new NotImplementedException();
         public Task<Result<Stream>> GetData() => throw new NotImplementedException();
+        public Task<Result> SetData(Stream stream, CancellationToken cancellationToken = default) => throw new NotImplementedException();
 
         public Task<Result> SetData(Stream data) => throw new NotImplementedException();
     }
