+++ test/Zafiro.FileSystem.Tests/ZafiroFileMixinTests.cs
@@ -35,6 +35,7 @@ public class TestZafiroFile : IZafiroFile
 
     public Task<Result> SetContents(IObservable<byte> contents, CancellationToken cancellationToken = default) => throw new NotImplementedException();
     public Task<Result<Stream>> GetData() => throw new NotImplementedException();
+    public Task<Result> SetData(Stream stream, CancellationToken cancellationToken = default) => throw new NotImplementedException();
 
     public Task<Result> SetData(Stream data) => throw new NotImplementedException();
 }
